Moved Networks class into its own file. - Removed import argparse ArgumentError.
